code refactoring; adding docs
