+++ task1/.gitignore
@@ -1,0 +1,3 @@
+gcs/*/output/
+app/*__pycache__
+*.venv/
