+++ task1/requirements.txt
@@ -1,0 +1,3 @@
+black==23.9.1
+pandas==2.0.1
+urllib3==1.26.18
